// frontend/src/gen/api/models/ProblemUpdate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * fk-mitou-2024
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 *
 * @export
 * @interface ProblemUpdate
 */
export interface ProblemUpdate {
  /**
   *
   * @type {string}
   * @memberof ProblemUpdate
   */
  description?: string | null;
  /**
   *
   * @type {boolean}
   * @memberof ProblemUpdate
   */
  isOpen?: boolean | null;
  /**
   *
   * @type {string}
   * @memberof ProblemUpdate
   */
  name?: string | null;
}

/**
 * Check if a given object implements the ProblemUpdate interface.
 */
export function instanceOfProblemUpdate(value: object): boolean {
  let isInstance = true;

  return isInstance;
}

export function ProblemUpdateFromJSON(json: any): ProblemUpdate {
  return ProblemUpdateFromJSONTyped(json, false);
}

export function ProblemUpdateFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): ProblemUpdate {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    description: !exists(json, 'description') ? undefined : json['description'],
    isOpen: !exists(json, 'is_open') ? undefined : json['is_open'],
    name: !exists(json, 'name') ? undefined : json['name'],
  };
}

export function ProblemUpdateToJSON(value?: ProblemUpdate | null): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    description: value.description,
    is_open: value.isOpen,
    name: value.name,
  };
}